01/11/2019

+++ Clinique/src/com/bridgeLabz/clinique/model/Patient.java
@@ -3,7 +3,7 @@ package com.bridgeLabz.clinique.model;
 public class Patient {
 	private String name;
 	private int id;
-	private int mobile_number;
+	private long mobile_number;
 	private int age;
 	public String getName() {
 		return name;
@@ -17,10 +17,10 @@ public class Patient {
 	public void setId(int id) {
 		this.id = id;
 	}
-	public int getMobile_number() {
+	public long getMobile_number() {
 		return mobile_number;
 	}
-	public void setMobile_number(int mobile_number) {
+	public void setMobile_number(long mobile_number) {
 		this.mobile_number = mobile_number;
 	}
 	public int getAge() {
